[linux] powermanagement: move implementation to platform folder
