+++ cmake/treedata/freebsd/subdirs.txt
@@ -3,10 +3,10 @@ xbmc/platform/linux                 platform/linux
 xbmc/platform/linux/input           platform/linux/input
 xbmc/platform/linux/network         platform/linux/network
 xbmc/platform/linux/peripherals     platform/linux/peripherals
+xbmc/platform/linux/powermanagement platform/linux/powermanagement
 xbmc/platform/linux/storage         platform/linux/storage
 xbmc/input/touch                    input/touch
 xbmc/input/touch/generic            input/touch/generic
-xbmc/powermanagement/linux          powermanagement/linux
 xbmc/filesystem/posix               filesystem/posix
 xbmc/utils/posix                    utils/posix
 xbmc/freebsd                        freebsdsupport
