+++ xbmc/platform/linux/powermanagement/ConsoleUPowerSyscall.cpp
