+++ xbmc/platform/linux/powermanagement/ConsoleUPowerSyscall.h
