+++ xbmc/platform/linux/powermanagement/FallbackPowerSyscall.h
