+++ xbmc/platform/linux/powermanagement/LinuxPowerSyscall.cpp
