+++ xbmc/platform/linux/powermanagement/LinuxPowerSyscall.h
