+++ xbmc/platform/linux/powermanagement/LogindUPowerSyscall.cpp
