+++ xbmc/platform/linux/powermanagement/LogindUPowerSyscall.h
