+++ xbmc/platform/linux/powermanagement/UPowerSyscall.cpp
