+++ xbmc/platform/linux/powermanagement/UPowerSyscall.h
