+++ xbmc/windowing/gbm/WinSystemGbm.cpp
@@ -27,7 +27,7 @@
 
 #include "OptionalsReg.h"
 #include "guilib/GraphicContext.h"
-#include "powermanagement/linux/LinuxPowerSyscall.h"
+#include "platform/linux/powermanagement/LinuxPowerSyscall.h"
 #include "settings/DisplaySettings.h"
 #include "utils/log.h"
 #include "utils/StringUtils.h"
