+++ xbmc/windowing/mir/WinSystemMir.cpp
@@ -24,7 +24,7 @@
 #include <string.h>
 
 #include "guilib/GraphicContext.h"
-#include "powermanagement/linux/LinuxPowerSyscall.h"
+#include "platform/linux/powermanagement/LinuxPowerSyscall.h"
 #include "settings/DisplaySettings.h"
 #include "utils/log.h"
 #include "WinEventsMir.h"
